+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Builtin documentation for builtin functions display from repl and command line (`musique doc <builtin>`)
+- Man documentation for commandline interface builtin (`musique man`)
+- Suggestions which command line parameters user may wanted to use
+
+### Changed
+
+- New parameter passing convention for command line invocation. `musique help` to learn how it changed
+
 ## [0.4.0]
 
 ### Added